:recycle: refactor routes structure to react router 6
