+++ client/src/scenes/App/App.jsx
@@ -1,35 +1,37 @@
-import React from 'react'
 import { Helmet } from 'react-helmet'
 
-import { ConfigurationProvider, AuthProvider, UserProvider } from 'contexts'
+import { AuthProvider, ConfigurationProvider, UserProvider } from 'contexts'
 
-import { AlertStack, AlertProvider } from 'components'
+import { AlertProvider, AlertStack } from 'components'
 
-import Navbar from './components/Navbar'
 import Footer from './components/Footer'
+import Navbar from './components/Navbar'
 
-import AppBody from './AppBody'
-
+import { Outlet } from 'react-router'
 import 'styles'
 
-const App = () => (
-  <div className="app theme">
-    <Helmet>
-      <title>FAQ</title>
-    </Helmet>
-    <AlertProvider>
-      <ConfigurationProvider>
-        <AuthProvider>
-          <UserProvider>
-            <Navbar />
-            <AppBody />
-            <Footer />
-          </UserProvider>
-        </AuthProvider>
-      </ConfigurationProvider>
-      <AlertStack />
-    </AlertProvider>
-  </div>
-)
+const App = () => {
+  return (
+    <div className="app theme">
+      <Helmet>
+        <title>FAQ</title>
+      </Helmet>
+      <AlertProvider>
+        <ConfigurationProvider>
+          <AuthProvider>
+            <UserProvider>
+              <Navbar />
+              <main className="main">
+                <Outlet />
+              </main>
+              <Footer />
+            </UserProvider>
+          </AuthProvider>
+        </ConfigurationProvider>
+        <AlertStack />
+      </AlertProvider>
+    </div>
+  )
+}
 
 export default App
